messaging refactor
